python/ai_server/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.32.0
pydantic==2.12.5
python-dotenv==1.0.1
requests==2.32.3
requests-ntlm==1.3.0
